requirements.txt
busylight-for-humans >= 0.7.6
Flask >= 1.1.2
flake8 >= 3.8.3
flask_restful >= 0.3.8
configparser >= 5.0.0